Rc to Box. Added Main(). Fixed dupes issue.
Didn't need to reference count the nodes of the trie, I just needed
them on the heap, so Box it is.  They still need to be refcounted
because they get modified during dictionary construction.

Discovered that if a boggle board legitimately had two ways of
generating the same word, that word would appear twice in the
solutions list.  That has been fixed.

Added very (VERY!) simple main() function.  It takes a boggle
board as a space-separated collection of letters and generates
all the solutions, printing the list of found words to the
console.