CLI is much more robust.
The command line tool now parses the board with more brains and better
failure modes.  It basically ignores whitespace, but does check to
make sure every row has the same number of letters.

An option to specify an alternative dictionary has been provided. A
copy of the unofficial Scrabble word dictionary for North America is
included in the root directory, heavily compressed.