FIX For Q.
Fixing the 'q' bug (damned English) was a mess and required a bit
of plumbing to (a) add 'u' to any 'q' instance on the board and (b)
convince the scanner that it was okay to use the same position twice,
but it worked.  It feels somewhat 1990s-ish, but it worked.

Yes, there's a unit test.  I worked hard to make a board that had a
lot of q words.