FEAT Multi-threaded Boggle solving!
There's probably a terrible performance hit for using the RWLock over
RefCell for the Trie, but it doesn't actually seem to be too bad in
practice.  Criterion says the "standard board" solver runes in about 1.07ms
for a single thread, 0.71 ms for two threads, and 0.54ms for four threads.
Those seem reasonable, and I'd be willing to bet some of the overhead is
the cost of building and destroying the board over and over to satisfy
the threading handler.