DOC The Prototype
This is the first draft of the Boggle solver, which I wrote in about
four hours after the original interview question.  It's... not too
shabby, actually.  It's in Python.  I'd *love* to be able to claim
that I wrote it in something awesomely functional like Hy, but I
struggled with that, so some vaguely imperative, vaguely OO code
will have to do.

This is the design that carried over so well into Rust, and you know,
it was easy to multi-thread, easy to adapt to my trie solution, and
easy to code.  I can't complain at all.